systemd: compile with BTF information
This adds the necessary BTF information for systemd's BPF programs. This relies on a purpose-built, minimal kernel, from which the `vmlinux.h` header file with the kernel-internal definitions is generated. Note that the systemd derivation is not depending on that specific kernel, it's just used for initial compilation. At runtime, the symbols are relocated using BPF CO-RE.